Add event listeners for functionality
A mouseover event listener was added to manipulate each square node in
the Nodelist, setting the color to green. The loop structure to create
the grid was moved into a playGame function along with the event for
the ability to call the function entirely when resetting the game with
the added "Click to Reset" button.

A prompt appears when the button is clicked and the input is used to
create a new grid to play the game once again.