Allow for grid editing and resetting
Function to add color to the grid was created upon mouseover. A button
was added to change the size of the grid in the reserved space as well
with a click event listener.

To be able to randomize the RGB values a "getRGB" function was created
and controlled with an if-else statement to maintain an initialized
color.